backend/requirements.txt
aiohttp
apprise
beautifulsoup4
fastapi
python-dotenv
uvicorn